requirements.txt
arxiv
requests
python-dotenv
paper-qa
fpdf
langchain
langchain-community
spacy>=3.7,<3.9
PyPDF2
openai